pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

# Python package metadata for the root integration package.
[project]
name = "vision_pipeline"
version = "0.1.0"
description = "Root integration environment for the vision pipeline."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "torch==2.6.0",
  "torchvision==0.21.0",
  "numpy==1.26.4",
  "hydra-core==1.3.2",
  "omegaconf==2.3.0",
  "iopath==0.1.10",
  "pillow==10.4.0",
  "opencv-python==4.10.0.84",
  "imageio==2.34.0",
  "imageio-ffmpeg==0.5.1",
  "scipy==1.12.0",
  "trimesh==4.2.2",
  "tqdm==4.66.5",
  "PyYAML==6.0.3",
  "click==8.1.8",
]

[tool.setuptools.packages.find]
include = ["src*"]

# ============================= Pixi workspace ============================= #

# The root workspace owns the shared ROS2 + segmentation integration envs.
[tool.pixi.workspace]
channels = ["https://prefix.dev/conda-forge"]
platforms = ["linux-64"]

# Source repository-local environment variables when entering `pixi shell`
# or running Pixi tasks. `env.sh` defines ROOT_DIR for path helpers.
[tool.pixi.activation]
scripts = ["env.sh"]

# ----------------------------- Base feature ------------------------------ #

# Packages shared by every root environment.
[tool.pixi.feature.base.dependencies]
python = "3.12.*"
numpy = "1.26.*"
pyyaml = "==6.0.3"
click = "8.1.*"
colcon-common-extensions = "*"
cmake = ">=3.28,<4"
ninja = ">=1.11,<2"
eigen = "3.4.*"

# Editable local packages and runtime-only PyPI dependencies.
[tool.pixi.feature.base.pypi-dependencies]
vision_pipeline = { path = ".", editable = true }
efficient_track_anything = { path = "segmentation/efficient_track_anything", editable = true }
sam2 = { path = "segmentation/sam2", editable = true }
pyrealsense2 = "==2.57.7.10387"

# ------------------------------ ROS2 features ----------------------------- #

# Default ROS2 stack.
[tool.pixi.feature.jazzy]
channels = ["https://prefix.dev/robostack-jazzy"]

[tool.pixi.feature.jazzy.dependencies]
ros-jazzy-ros-base = "*"
ros-jazzy-sensor-msgs = "*"
ros-jazzy-geometry-msgs = "*"
ros-jazzy-rviz2 = "*"

# Alternate ROS2 stack for Humble systems.
[tool.pixi.feature.humble]
channels = ["https://prefix.dev/robostack-humble"]

[tool.pixi.feature.humble.dependencies]
ros-humble-ros-base = "*"
ros-humble-sensor-msgs = "*"
ros-humble-geometry-msgs = "*"
ros-humble-rviz2 = "*"

# ------------------------------ CUDA features ----------------------------- #

# CUDA 12.6 build toolkit for the default Jazzy workflow.
[tool.pixi.feature.cuda126.dependencies]
cuda-nvcc_linux-64 = "12.6.*"
cuda-libraries-dev = "12.6.*"

[tool.pixi.feature.cuda126.system-requirements]
cuda = "12.6"

# CUDA 12.4 build toolkit for the Humble workflow.
[tool.pixi.feature.cuda124.dependencies]
cuda-nvcc_linux-64 = "12.4.*"
cuda-libraries-dev = "12.4.*"

[tool.pixi.feature.cuda124.system-requirements]
cuda = "12.4"

# ------------------------------ Environments ------------------------------ #

# Keep environments explicit so task defaults and manual `-e ...` selection are obvious.
[tool.pixi.environments]
default = { features = ["base", "jazzy", "cuda126"], solve-group = "jazzy-cuda126" }
jazzy-cuda126 = { features = ["base", "jazzy", "cuda126"], solve-group = "jazzy-cuda126" }
humble-cuda124 = { features = ["base", "humble", "cuda124"], solve-group = "humble-cuda124" }

# --------------------------------- Tasks ---------------------------------- #

# Operational helpers: download segmentation checkpoints and switch CUDA build toolkits.
[tool.pixi.tasks]

[tool.pixi.tasks.download-seg-ckpt]
cmd = "bash checkpoints/download_ckpts.sh"
description = "Download the default SAM2 and EfficientTAM segmentation checkpoints."

[tool.pixi.tasks.download_seg-ckpt-all]
cmd = "bash checkpoints/download_ckpts_all.sh"
description = "Download all SAM2, SAM2.1, and EfficientTAM segmentation checkpoints."

[tool.pixi.tasks.ros2-tracking-build]
cmd = "colcon --log-base ros2/tracking/log build --base-paths ros2/tracking --packages-select tracking --build-base ros2/tracking/build --install-base ros2/tracking/install"
description = "Build the ROS2 tracking package with package-local build, install, and log directories."
default-environment = "jazzy-cuda126"

[tool.pixi.tasks.rviz]
cmd = "rviz2"
description = "Launch RViz2 from the default Jazzy ROS2 environment."
default-environment = "jazzy-cuda126"

[tool.pixi.tasks.cuda126-shell]
cmd = '''
bash -lc 'cuda_home=$(dirname "$(dirname "$(command -v nvcc)")"); export CUDA_HOME="$cuda_home"; export CUDA_PATH="$cuda_home"; export CUDACXX="$cuda_home/bin/nvcc"; echo "Using CUDA toolkit at $CUDA_HOME"; exec bash'
'''
description = "Open a shell configured for the Jazzy CUDA 12.6 build toolkit."
default-environment = "jazzy-cuda126"

[tool.pixi.tasks.cuda124-shell]
cmd = '''
bash -lc 'cuda_home=$(dirname "$(dirname "$(command -v nvcc)")"); export CUDA_HOME="$cuda_home"; export CUDA_PATH="$cuda_home"; export CUDACXX="$cuda_home/bin/nvcc"; echo "Using CUDA toolkit at $CUDA_HOME"; exec bash'
'''
description = "Open a shell configured for the Humble CUDA 12.4 build toolkit."
default-environment = "humble-cuda124"
